tidy up lib package, move classes in own packages
