+++ app/src/main/java/vocabletrainer/heinecke/aron/vocabletrainer/lib/CSV/CSVHeaders.java
@@ -1,4 +1,4 @@
-package vocabletrainer.heinecke.aron.vocabletrainer.lib;
+package vocabletrainer.heinecke.aron.vocabletrainer.lib.CSV;
 
 /**
  * Class for header commands
