+++ app/src/main/java/vocabletrainer/heinecke/aron/vocabletrainer/lib/CSV/Import/Importer.java
@@ -1,4 +1,4 @@
-package vocabletrainer.heinecke.aron.vocabletrainer.lib.Import;
+package vocabletrainer.heinecke.aron.vocabletrainer.lib.CSV.Import;
 
 import android.content.Context;
 import android.util.Log;
