+++ app/src/main/java/vocabletrainer/heinecke/aron/vocabletrainer/lib/CSV/Import/PreviewParser.java
@@ -1,4 +1,4 @@
-package vocabletrainer.heinecke.aron.vocabletrainer.lib.Import;
+package vocabletrainer.heinecke.aron.vocabletrainer.lib.CSV.Import;
 
 import java.util.ArrayList;
 import java.util.List;
