+++ app/src/main/java/vocabletrainer/heinecke/aron/vocabletrainer/lib/Widget/DPIHelper.java
@@ -1,4 +1,4 @@
-package vocabletrainer.heinecke.aron.vocabletrainer.lib;
+package vocabletrainer.heinecke.aron.vocabletrainer.lib.Widget;
 
 import android.content.res.Resources;
 import android.util.TypedValue;
