+++ app/src/main/res/layout/fragment_trainer_classic.xml
@@ -25,7 +25,7 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.0">
 
-        <vocabletrainer.heinecke.aron.vocabletrainer.lib.CEditText
+        <vocabletrainer.heinecke.aron.vocabletrainer.lib.Widget.CEditText
             android:id="@+id/tTrainerInput"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
