+++ app/src/test/java/vocabletrainer/heinecke/aron/vocabletrainer/MultiMeaningParserTest.java
@@ -7,8 +7,8 @@ import org.junit.Test;
 import java.util.Arrays;
 import java.util.List;
 
-import vocabletrainer.heinecke.aron.vocabletrainer.lib.CSVCustomFormat;
-import vocabletrainer.heinecke.aron.vocabletrainer.lib.MultiMeaningHandler;
+import vocabletrainer.heinecke.aron.vocabletrainer.lib.CSV.CSVCustomFormat;
+import vocabletrainer.heinecke.aron.vocabletrainer.lib.CSV.MultiMeaningHandler;
 
 import static org.junit.Assert.*;
